Added fields to parts, added some part of configurator module

+++ src/main/java/com/wisepc/wisepc/model/parts/domain/Processor.java
@@ -26,5 +26,5 @@ public class Processor extends PartEntity {
     private Double frequency;
 
     @Column(name = "cores")
-    private Long cores;
+    private Integer cores;
 }
